make bu_bitv_shift() return a size_t instead of an unsigned int in order to avoid a type downcast warning on windows
svn:revision:55411
svn:branch:trunk
svn:account:brlcad

+++ include/bu.h
@@ -1412,7 +1412,7 @@ typedef struct bu_bitv bu_bitv_t;
  * length sizeof(bitv_t)*8.0 bits long.  users should not call this
  * directly, instead calling the BU_BITV_SHIFT macro instead.
  */
-BU_EXPORT extern unsigned int bu_bitv_shift();
+BU_EXPORT extern size_t bu_bitv_shift();
 
 /*
  * Bit-string manipulators for arbitrarily long bit strings stored as
